Added tests.

+++ .eng/shared/build/Build.ps1
@@ -38,7 +38,10 @@ param (
 [switch] $Prepare = $false,
 
 # Runs the test suite.
-[switch] $Test = $false
+[switch] $Test = $false,
+
+# Verifies the test coverage (requires -Test).
+[switch] $Coverage = $false
 
 )
 
@@ -153,21 +156,27 @@ function Sign() {
 function Test() {
      $testResultsDir = $(Get-Location).Path + "\TestResults"
 
-    # Removing the TestResults directory so that we reset the code coverage information.
-    if ( Test-Path $testResultsDir ) {
-        Remove-Item $testResultsDir -Recurse -Force
+    if ( $Coverage ) {
+         # Removing the TestResults directory so that we reset the code coverage information.
+        if ( Test-Path $testResultsDir ) {
+            Remove-Item $testResultsDir -Recurse -Force
+        }
+    
+        # Restoring the required dotnet tools.
+        & $PSScriptRoot/../RestoreTools.ps1 PostSharp.Engineering.BuildTools
+    
+        # Executing tests with code coverage enabled.
+        & dotnet test -p:CollectCoverage=True -p:CoverletOutput="$testResultsDir\" -m:1 --nologo --no-restore
+        if ($LASTEXITCODE -ne 0 ) { throw "Tests failed." }
+    
+        # Detect gaps in code coverage.
+        & tools/postsharp-eng.exe coverage warn "$testResultsDir\coverage.opencover.xml"
+        if ($LASTEXITCODE -ne 0 ) { throw "Test coverage has gaps." }
+    } else {
+        # Executing tests without test coverage
+        & dotnet test --nologo --no-restore
+        if ($LASTEXITCODE -ne 0 ) { throw "Tests failed." }
     }
-
-    # Restoring the required dotnet tools.
-    & $PSScriptRoot/../RestoreTools.ps1 PostSharp.Engineering.BuildTools
-
-    # Executing tests with code coverage enabled.
-    & dotnet test -p:CollectCoverage=True -p:CoverletOutput="$testResultsDir\" -m:1 --nologo --no-restore
-    if ($LASTEXITCODE -ne 0 ) { throw "Tests failed." }
-
-    # Detect gaps in code coverage.
-    & tools/postsharp-eng.exe coverage warn "$testResultsDir\coverage.opencover.xml"
-    if ($LASTEXITCODE -ne 0 ) { throw "Test coverage has gaps." }
 
 
     Write-Host "Tests successful" -ForegroundColor Green
